Make all actions work

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/messages/ColorMessage.kt
@@ -4,10 +4,10 @@ import com.qualcomm.robotcore.hardware.NormalizedRGBA
 
 class ColorMessage(color: NormalizedRGBA) {
     @JvmField val timestamp = System.nanoTime();
-    @JvmField val r = color.red;
-    @JvmField val g = color.green;
-    @JvmField val b = color.blue;
-    @JvmField val a = color.alpha;
+    @JvmField val r = color.red.toDouble();
+    @JvmField val g = color.green.toDouble();
+    @JvmField val b = color.blue.toDouble();
+    @JvmField val a = color.alpha.toDouble();
 
     override fun toString(): String {
         return "$r $g $b $a"
